+++ attributes/default.rb
@@ -1,3 +1,3 @@
 # encoding: UTF-8
 
-default['partition_resize']['compiletime'] = false
+default['fs_resize']['compiletime'] = false