+++ metadata.rb
@@ -1,11 +1,12 @@
 # encoding: UTF-8
 
-name 'partition_resize'
+name 'fs_resize'
 maintainer 'Onddo Labs, Sl.'
 maintainer_email 'team@onddo.com'
 license 'Apache 2.0'
 description <<-EOS
-Resize partitions automatically when the underlying disk increases its size.
+Resize the file system automatically when the underlying partition or disk
+increases its size.
 EOS
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version '0.1.0'
@@ -17,11 +18,11 @@ supports 'fedora'
 supports 'redhat'
 supports 'ubuntu', '>= 12.04'
 
-recipe 'partition_resize::default', 'Resize mounted partitions.'
+recipe 'fs_resize::default', 'Resize mounted file systems.'
 
-attribute 'partition_resize/compiletime',
-          display_name: 'partition resize at compile time',
-          description: 'Resize the partitions at compile time.',
+attribute 'fs_resize/compiletime',
+          display_name: 'fs resize at compile time',
+          description: 'Resize the file sistems at compile time.',
           choice: %w(true false),
           type: 'string',
           required: 'optional',