+++ test/kitchen/cookbooks/fs_resize-test/CHANGELOG.md
@@ -1,11 +1,11 @@
-partition_resize-test CHANGELOG
+fs_resize-test CHANGELOG
 ===============================
 
-This file is used to list changes made in each version of the partition_resize-test cookbook.
+This file is used to list changes made in each version of the fs_resize-test cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of partition_resize-test
+- [your_name] - Initial release of fs_resize-test
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.